Protect admin and gateone web sites behind basic authentication backed by LDAP.
This is made possible by configuring nginx to delegate this basic
authentication to a proxy HTTP server, managed by a new service
nginx-ldap-auth-daemon, and which uses Python's LDAP APIs to check the
credentials.

This is based on Nginx Inc.'s reference implementation of an
authentication system. See the following links for more information:

 * https://github.com/nginxinc/nginx-ldap-auth
 * https://www.nginx.com/blog/nginx-plus-authenticate-users

Aside from having largely rewritten nginx-ldap-auth-daemon's code to
better match my personal coding style, I also added the ability to
check for the authenticated user's membership in a certain group.

Additionally, I replaced the extensive command line interface with a
simple configuration file; not only can that be pretty convenient to
set once and for all configuration keys with default values for any site
protected behind LDAP authentication (although all those things can be
overridden on a per-site basis with the X-LDAP proxy headers).
One of the nicest things about the configuration file is that, if all
the sites protected behind LDAP authentication tap into the same LDAP
server, the URL and the bind credentials can be stored in that file, and
thus won't show up in the output of `ps aux`, and the file can be
protected with appropriate file permissions.