Minor changes to zsh configuration around places where we check the presence or path of an executable using which.
That tool, in zsh, is a builtin that isn't quite as fully featured as
the actual binary in /usr/bin/which. Instead, we can use the commands
array in zsh to either check for the presence of a command without
needing to pipe the result to /dev/null, or get the full path to the
command. Additionally, in the few places where we are currently
implementing a function with the same name of an external command, to
enhance it, it's much simpler to use `command
<name_of_the_external_command>`; besides that, this allows us to avoid
hardcoding the path of the underlying binary, which is a nice added
benefit.