Improve LDAP directory with two new auxiliary object classes.
 * `fullyQualifiedGroup`: in the context of this sort of multi-domain
directory, we may end up with multiple groups in the various domain that
have the same value for the cn attribute; to better handle such possible
collisions, it may be useful to have a new field, which we'll call
`fullyQualifiedGroupName`, to disambiguate between such groups; for
example, we may simply fill this attribute with the concatenation of the
group's common name and the domain component, separated by some special
character, like the `%` sign.

 * `applicationParticipant`: this adds a new attribute,
`supportedApplication`, which can be appended to users or groups, one or
more times, in order to denote which entities should be opted into
being integrated with certain applications.