Add configuration to support virtual mail authenticated senders without a virtual mailbox, in order to allow sending messages from an email address whose domain is properly configured to take advantage of DKIM and SPF.
This was needed for my new multi-function printer to be able to send
scanned documents via email. This is the best way I found to lower the
chances of those emails being flagged as spam, despite the mail client
of the printer being far from optimal (most notably with a bad HELO
command).

Lastly, the printer wasn't consistently including a Message-ID header,
so it may help to enable the `always_add_missing_headers` option in
postfix.