Adjustments to rspamd antivirus configuration to guarantee proper scanning of messages.
I was experiencing a problem similar to the one described by @ygoe in
this issue: https://github.com/rspamd/rspamd/issues/3503#issuecomment-702657567

Specifically, when sending the EICAR virus test email in a quick and
dirty manner from the command line, piping echo and cat to sendmail, the
incoming message was never even sent to ClamAV for scanning.

Like @ygoe, I found that the same email would be rejected just fine
after passing through ClamAV if `scan_mime_parts` is set to `false`, to
send the entire message to ClamAV.

However, similar messages sent from the same host with a more
conventional method, i.e. using a proper mail client like mutt, would
be sent to ClamAV just fine for scanning with `scan_mime_parts` set to
`true`.

So, in my opinion, the best way to guarantee that all the possible ways
to send a virus (or the EICAR antivirus test file) would be detected by
rspamd as a virus is to use two separate antivirus scanners with
different settings as to how incoming messages are supposed to be scanned.
The first scanner will only scan the entire message, with
`scan_mime_parts` set to `false`, while the second scanner will scan
MIME parts separately, with `scan_mime_parts` set to `true`.