Add per-user bayes classification statistics for rspamd.
rspamd's documentation suggests to use LDA mode in order to avoid
issues with per-user statistics when confronted with multi-recipient
messages.

However, that suggestion has significant drawbacks, since it makes
difficult to apply certain actions, such as greylisting.

So rather than losing on the benefits of a simple configuration for
postfix to talk to rspamd using the milter infrastructure, this approach
just adds a new step before local delivery of mail by having incoming
mail be relayed through a new content-filter, defined in postfix's
master.cf, named `sendmail`, configured with a destination recipient
limit of 1.

So in the case of an incoming message intended for n local mailbox
users, the incoming mail will be relayed n times for each of the local
mailbox users; hence, it will go through the milter (i.e. rspamd) a
total of n + 1 times: once for the incoming connection from the
originating external mail server, and once for each of the local mailbox
users.

The `sendmail` content-filter is configured to add the "X-Original-To"
header, which can then be used to distinguish rspamd passes for the
incoming connection from the originating mail server on the one hand,
from passes for the delivery to each of the local mailbox recipients on
the other hand.

In the first rspamd pass for the incoming connection from the
originating mail server learning for the Bayes per-user classifier is
restricted, such that it is only enabled for the Bayes global
classifier, whereas the exact opposite happens for passes for the
delivery to each of the local mailbox recipients.

This kind of setup might warrant tweaking scores for the BAYES_HAM and
BAYES_SPAM symbols, which are still used for the Bayes global
classifier, as their presence could fight with a contradictory per-user
Bayes symbol, but more testing would be required to have a reasonable
level of confidence about such a change.